# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Incidents",
        "description": "API complète pour la gestion des incidents municipaux",
        "contact": {
            "name": "Support API",
            "email": "support@incidents.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/conversations": {
            "get": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Mes conversations",
                "operationId": "c6e2de76967e6131ee742d053dec11a7",
                "responses": {
                    "200": {
                        "description": "Liste des conversations"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Démarrer une conversation (Citizen)",
                "operationId": "58fb50267e9f85b4ed687e3182f6d7ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "admin_id",
                                    "message"
                                ],
                                "properties": {
                                    "admin_id": {
                                        "description": "ID du responsable municipal",
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Bonjour, je souhaite signaler un problème"
                                    },
                                    "incident_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conversation démarrée"
                    },
                    "404": {
                        "description": "Responsable municipal non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/conversations/{id}": {
            "get": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Voir une conversation avec ses messages",
                "operationId": "ddbe0a94327643b2ce46ae01a43da3ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation avec messages"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Conversation non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/conversations/{id}/messages": {
            "post": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Envoyer un message",
                "operationId": "1804c269b4621f20db484a821f90d575",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Merci pour votre retour"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message envoyé"
                    },
                    "400": {
                        "description": "Conversation fermée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/conversations/{id}/close": {
            "patch": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Fermer une conversation (Admin municipal)",
                "operationId": "630b1687055cc27e3b0fa6ff14eeae31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation fermée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Conversation non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/messages/{id}": {
            "delete": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Supprimer un message",
                "operationId": "b48b82c8aad44f54e8eb7cbc2d7f298f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message supprimé"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum (format : Bearer <token>)"
            }
        }
    },
...
}